updating product category
