+++ blocks/product-category/product-category.js
@@ -2,7 +2,9 @@ import { div, span, a, img, p } from "../../scripts/dom-builder.js";
 import { getProductsForCategories } from "../../scripts/commerce.js";
 import { makePublicUrl, imageHelper } from "../../scripts/scripts.js";
 import { decorateIcons } from "../../scripts/lib-franklin.js";
-
+import renderGridCard from "./gridData.js";
+import renderListCard from "./listData.js";
+import renderPagination from "./pagination.js";
 async function fetchProducts() {
   try {
     const productsCategories = await getProductsForCategories();
@@ -13,11 +15,13 @@ async function fetchProducts() {
   }
 }
 
+console.log("renderGridCard", renderGridCard);
+console.log("renderListCard", renderListCard);
+
 export default async function decorate(block) {
   const content = block.querySelector("div");
   const productsCategories = await fetchProducts();
 
-  // Fixed cards per page to 3 as requested
   const cardsPerPage = 4;
   const cardsPerPageList = 3;
   let currentPage = 1;
@@ -109,561 +113,6 @@ export default async function decorate(block) {
     class:
       "pagination-container flex justify-center items-center gap-2 mt-8 w-full",
   });
-
-  function renderGridCard(item) {
-    const card = div({
-      class:
-        "w-full sm:w-[calc(50%-10px)] lg:w-[calc(25%-15px)] min-h-80 bg-white outline outline-1 outline-gray-300 flex flex-col justify-start items-start",
-    });
-
-    const image = imageHelper(item.raw.images[0], item.title, {
-      href: makePublicUrl(item.path),
-      title: item.title,
-      class: "w-full h-40 object-cover",
-    });
-
-    const head = p(
-      { class: "p-3 text-black text-xl font-normal leading-7" },
-      item.title
-    );
-    const desc = p(
-      { class: "p-3 text-gray-700 text-base font-extralight leading-snug" },
-      item?.raw?.source
-    );
-
-    const details = div(
-      {
-        class:
-          "self-stretch px-4 py-3 bg-gray-50 inline-flex flex-col justify-start items-end gap-6",
-      },
-      div(
-        {
-          class:
-            "text-right justify-start text-black text-2xl font-normal font-['TWK_Lausanne_Pan'] leading-loose",
-        },
-        "$1,000.00"
-      ),
-      div(
-        { class: "self-stretch flex flex-col justify-start items-start gap-2" },
-        div(
-          { class: "flex justify-between items-center w-full" },
-          div(
-            {
-              class:
-                "text-black text-base font-extralight font-['TWK_Lausanne_Pan'] leading-snug",
-            },
-            "Unit of Measure:"
-          ),
-          div(
-            {
-              class:
-                "text-black text-base font-bold font-['TWK_Lausanne_Pan'] leading-snug",
-            },
-            item?.raw?.uom || "1/Bundle"
-          )
-        ),
-        div(
-          { class: "flex justify-between items-center w-full" },
-          div(
-            {
-              class:
-                "text-black text-base font-extralight font-['TWK_Lausanne_Pan'] leading-snug",
-            },
-            "Min. Order Qty:"
-          ),
-          div(
-            {
-              class:
-                "text-black text-base font-bold font-['TWK_Lausanne_Pan'] leading-snug",
-            },
-            item?.raw?.minQty || "50"
-          )
-        )
-      ),
-      div(
-        { class: "inline-flex justify-start items-center gap-3" },
-        div(
-          {
-            class:
-              "w-14 self-stretch px-4 py-1.5 bg-white rounded-md shadow-sm outline outline-1 outline-offset-[-1px] outline-gray-300 flex justify-center items-center overflow-hidden",
-          },
-          div(
-            {
-              class:
-                "justify-start text-black text-base font-normal font-['Inter'] leading-normal",
-            },
-            "1"
-          )
-        ),
-        div(
-          {
-            class:
-              "w-24 px-5 py-2 bg-violet-600 rounded-[20px] outline outline-1 outline-offset-[-1px] outline-violet-600 flex justify-center items-center overflow-hidden",
-          },
-          div(
-            {
-              class:
-                "text-white text-base font-normal font-['TWK_Lausanne_Pan'] leading-snug",
-            },
-            "Buy"
-          )
-        ),
-        div(
-          {
-            class:
-              "px-5 py-2 bg-white rounded-[20px] outline outline-1 outline-offset-[-1px] outline-violet-600 flex justify-center items-center overflow-hidden",
-          },
-          div(
-            {
-              class:
-                "text-violet-600 text-base font-normal font-['TWK_Lausanne_Pan'] leading-snug",
-            },
-            "Quote"
-          )
-        )
-      )
-    );
-
-    const spacer = div({ class: "flex-grow" });
-    const desc1 = div(
-      { class: "self-stretch p-3 flex justify-start items-center" },
-      div(
-        { class: "text-violet-600 text-base font-bold leading-snug" },
-        "View Details →"
-      )
-    );
-
-    card.append(image, head, desc, details, spacer, desc1);
-    return card;
-  }
-
-  function renderListCard(item) {
-    return div(
-      {
-        class:
-          "self-stretch w-full outline outline-1 outline-gray-300 inline-flex justify-start items-center flex-wrap md:flex-nowrap",
-      },
-      div(
-        {
-          class:
-            "flex-1 self-stretch p-6 bg-white flex justify-start items-start gap-6",
-        },
-        div(
-          {
-            class: "w-24 inline-flex flex-col justify-start items-center gap-3",
-          },
-          div(
-            {
-              class:
-                "self-stretch max-h-24 relative rounded-md outline outline-1 outline-offset-[-1px] outline-gray-900",
-            },
-            div({
-              class: "w-24 h-24 left-0 top-0 absolute bg-white rounded-md",
-            }),
-            img({
-              class:
-                "w-24 h-24 left-0 top-0 absolute rounded-md border border-gray-200 object-cover",
-              src: item?.raw?.images?.[0] || "",
-              alt: item.title || "",
-            })
-          )
-        ),
-        div(
-          {
-            class:
-              "self-stretch h-44 inline-flex flex-col justify-between items-start",
-          },
-          div(
-            {
-              class:
-                "self-stretch flex flex-col justify-start items-start gap-3",
-            },
-            div(
-              {
-                class:
-                  "self-stretch flex flex-col justify-start items-start gap-1",
-              },
-              item?.raw?.tag &&
-                div(
-                  {
-                    class:
-                      "px-4 py-1 bg-violet-50 inline-flex justify-center items-center gap-2.5",
-                  },
-                  div(
-                    {
-                      class:
-                        "text-center justify-start text-violet-600 text-sm font-normal font-['TWK_Lausanne_Pan'] leading-tight",
-                    },
-                    item.raw.tag
-                  )
-                ),
-              div(
-                {
-                  class:
-                    "self-stretch justify-start text-black text-xl font-normal font-['TWK_Lausanne_Pan'] leading-7",
-                },
-                item.title
-              )
-            ),
-            div(
-              {
-                class:
-                  "self-stretch inline-flex justify-start items-center gap-3",
-              },
-              div(
-                {
-                  class:
-                    "flex-1 inline-flex flex-col justify-start items-start",
-                },
-                div(
-                  {
-                    class:
-                      "w-64 justify-start text-gray-700 text-base font-extralight font-['TWK_Lausanne_Pan'] leading-snug",
-                  },
-                  item?.raw?.sku || ""
-                )
-              )
-            ),
-            // Adding View Details and Source below SKU only in desktop view
-            div(
-              {
-                class: "hidden md:flex w-full flex-col gap-2 mt-4", // This section is only visible in desktop view
-              },
-              div(
-                {
-                  class: "text-left text-violet-600 font-bold",
-                },
-                "View Details →"
-              ),
-              div(
-                {
-                  class: "text-left text-sm text-gray-600",
-                },
-                item?.raw?.source || "No Source Available"
-              )
-            )
-          )
-        )
-      ),
-      // Mobile-specific: New row for "View Details" and "Source" - placed below image and product details for mobile view
-      div(
-        {
-          class: "w-full flex flex-wrap md:hidden", // Only show in mobile (md:hidden hides this on desktop)
-        },
-        div(
-          {
-            class: "text-left text-violet-600 font-bold w-full", // Stack on mobile, align left
-          },
-          "View Details →"
-        ),
-        div(
-          {
-            class: "text-left text-sm text-gray-600 w-full", // Stack on mobile, align left
-          },
-          item?.raw?.source || "No Source Available"
-        )
-      ),
-      div(
-        {
-          class:
-            "self-stretch p-6 bg-gray-50 inline-flex flex-col justify-start items-end gap-4",
-        },
-        div(
-          {
-            class:
-              "w-64 text-right justify-start text-black text-2xl font-normal font-['TWK_Lausanne_Pan'] leading-loose",
-          },
-          `$${item?.raw?.price || "50.00"}`
-        ),
-        div(
-          {
-            class: "w-64 flex flex-col gap-2",
-          },
-          div(
-            {
-              class: "flex justify-between items-center",
-            },
-            div(
-              {
-                class:
-                  "text-black text-base font-extralight font-['TWK_Lausanne_Pan'] leading-snug",
-              },
-              "Availability:"
-            ),
-            div(
-              {
-                class:
-                  "text-black text-base font-bold font-['TWK_Lausanne_Pan'] leading-snug",
-              },
-              `${item?.raw?.availability || 0} Available`
-            )
-          ),
-          div(
-            {
-              class: "flex justify-between items-center",
-            },
-            div(
-              {
-                class:
-                  "text-black text-base font-extralight font-['TWK_Lausanne_Pan'] leading-snug",
-              },
-              "Unit of Measure:"
-            ),
-            div(
-              {
-                class:
-                  "text-black text-base font-bold font-['TWK_Lausanne_Pan'] leading-snug",
-              },
-              item?.raw?.uom || "2/Bundle"
-            )
-          ),
-          div(
-            {
-              class: "flex justify-between items-center",
-            },
-            div(
-              {
-                class:
-                  "text-black text-base font-extralight font-['TWK_Lausanne_Pan'] leading-snug",
-              },
-              "Min. Order Qty:"
-            ),
-            div(
-              {
-                class:
-                  "text-black text-base font-bold font-['TWK_Lausanne_Pan'] leading-snug",
-              },
-              item?.raw?.minQty || "4"
-            )
-          )
-        ),
-        div(
-          {
-            class: "inline-flex justify-start items-center gap-3",
-          },
-          div(
-            {
-              class:
-                "w-14 self-stretch px-4 py-1.5 bg-white rounded-md shadow-[0px_1px_2px_0px_rgba(0,0,0,0.05)] outline outline-1 outline-offset-[-1px] outline-gray-300 flex justify-center items-center overflow-hidden",
-            },
-            div(
-              {
-                class:
-                  "justify-start text-black text-base font-normal font-['Inter'] leading-normal",
-              },
-              "1"
-            )
-          ),
-          div(
-            {
-              class:
-                "w-24 px-5 py-2 bg-violet-600 rounded-[20px] outline outline-1 outline-offset-[-1px] outline-violet-600 flex justify-center items-center overflow-hidden",
-            },
-            div(
-              {
-                class:
-                  "justify-start text-white text-base font-normal font-['TWK_Lausanne_Pan'] leading-snug",
-              },
-              "Buy"
-            )
-          ),
-          div(
-            {
-              class:
-                "px-5 py-2 bg-white rounded-[20px] outline outline-1 outline-offset-[-1px] outline-violet-600 flex justify-center items-center overflow-hidden",
-            },
-            div(
-              {
-                class:
-                  "justify-start text-violet-600 text-base font-normal font-['TWK_Lausanne_Pan'] leading-snug",
-              },
-              "Quote"
-            )
-          )
-        )
-      )
-    );
-  }
-
-  function renderPagination() {
-    paginationContainer.innerHTML = "";
-
-    const totalPages = Math.ceil(productsCategories.length / cardsPerPageList);
-
-    // Create the container for pagination with left-aligned and right-aligned buttons
-    const paginationWrapper = div({
-      class: "inline-flex w-full items-center justify-between", // Aligns Previous and Next buttons on the left and right
-    });
-
-    // Previous button (left-aligned)
-    const prevButton = div(
-      {
-        class: `flex items-center gap-1 cursor-pointer ${
-          currentPage === 1
-            ? "text-gray-400 cursor-not-allowed"
-            : "text-violet-600 hover:underline"
-        }`,
-      },
-      div(
-        { class: "w-5 h-5 relative overflow-hidden" },
-        span({
-          class: `icon icon-arrow-left w-6 h-6 absolute fill-current ${
-            currentPage === 1 ? "text-gray-400" : "text-violet-600"
-          } [&_svg>use]:stroke-current`,
-        })
-      ),
-      span(
-        { class: `${currentPage === 1 ? "text-gray-400" : "text-violet-600"}` },
-        "Previous"
-      ) // Label
-    );
-
-    decorateIcons(prevButton);
-
-    prevButton.addEventListener("click", () => {
-      if (currentPage > 1) {
-        currentPage--;
-        updateCarousel();
-      }
-    });
-
-    // Page numbers container - centered
-    const pageNumbersContainer = div({
-      class: "flex items-center justify-center gap-1", // Reduced gap between page numbers
-    });
-
-    // Page numbers logic
-    const maxVisiblePages = 5;
-    let startPage = Math.max(1, currentPage - Math.floor(maxVisiblePages / 2));
-    const endPage = Math.min(totalPages, startPage + maxVisiblePages - 1);
-
-    if (endPage - startPage + 1 < maxVisiblePages) {
-      startPage = Math.max(1, endPage - maxVisiblePages + 1);
-    }
-
-    // First page
-    if (startPage > 1) {
-      const firstPage = div(
-        {
-          class: `w-8 h-8 flex items-center justify-center rounded-md cursor-pointer ${
-            currentPage === 1 ? "bg-violet-600 text-white" : "hover:bg-gray-100"
-          }`,
-        },
-        "1"
-      );
-
-      firstPage.addEventListener("click", () => {
-        currentPage = 1;
-        updateCarousel();
-      });
-
-      pageNumbersContainer.append(firstPage);
-
-      // Ellipsis if needed
-      if (startPage > 2) {
-        const ellipsis = div(
-          {
-            class: "w-8 h-8 flex items-center justify-center",
-          },
-          "..."
-        );
-        pageNumbersContainer.append(ellipsis);
-      }
-    }
-
-    // Page numbers
-    for (let i = startPage; i <= endPage; i++) {
-      const pageNumber = div(
-        {
-          class: `w-8 h-8 flex items-center justify-center rounded-md cursor-pointer ${
-            currentPage === i ? "bg-violet-600 text-white" : "hover:bg-gray-100"
-          }`,
-        },
-        i.toString()
-      );
-
-      pageNumber.addEventListener("click", () => {
-        currentPage = i;
-        updateCarousel();
-      });
-
-      pageNumbersContainer.append(pageNumber);
-    }
-
-    // Ellipsis if needed
-    if (endPage < totalPages - 1) {
-      const ellipsis = div(
-        {
-          class: "w-8 h-8 flex items-center justify-center",
-        },
-        "..."
-      );
-      pageNumbersContainer.append(ellipsis);
-    }
-
-    // Last page if not already included
-    if (endPage < totalPages) {
-      const lastPage = div(
-        {
-          class: `w-8 h-8 flex items-center justify-center rounded-md cursor-pointer ${
-            currentPage === totalPages
-              ? "bg-violet-600 text-white"
-              : "hover:bg-gray-100"
-          }`,
-        },
-        totalPages.toString()
-      );
-
-      lastPage.addEventListener("click", () => {
-        currentPage = totalPages;
-        updateCarousel();
-      });
-
-      pageNumbersContainer.append(lastPage);
-    }
-
-    // Next button (right-aligned)
-    const nextButton = div(
-      {
-        class: `flex mr-2 items-center cursor-pointer ${
-          currentPage === totalPages
-            ? "text-gray-400 cursor-not-allowed"
-            : "text-violet-600 hover:underline"
-        }`,
-      },
-      span(
-        {
-          class: `${
-            currentPage === totalPages ? "text-gray-400" : "text-violet-600"
-          }`, // Change color based on the functionality
-        },
-        "Next"
-      ), // Label
-      div(
-        {
-          class: "w-5 h-5 relative overflow-hidden",
-        },
-        span({
-          class: `icon icon-arrow-right w-6 h-6 absolute fill-current ${
-            currentPage === totalPages ? "text-gray-400" : "text-violet-600"
-          } [&_svg>use]:stroke-current`,
-        })
-      )
-    );
-
-    decorateIcons(nextButton);
-
-    nextButton.addEventListener("click", () => {
-      if (currentPage < totalPages) {
-        currentPage++;
-        updateCarousel();
-      }
-    });
-
-    // Append everything into the container
-    paginationWrapper.append(prevButton, pageNumbersContainer, nextButton);
-    paginationContainer.append(paginationWrapper);
-  }
 
   function updateCarousel() {
     carouselCards.innerHTML = "";
